ServerFinder: Addressed a potential race condition.
It happened in the past that responses sent by the server were missed.
This was probably caused by the receiver not yet ready. To fix this,
sender and receiver now run in different coroutines, and the sender
sends requests periodically. Thus, if one request is missed, it is
likely that the next one is received.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>