app: The view model is now injected by Hilt.
For this to work, the `ControlViewModel` interface had to be converted
to an abstract class, since a view model must extend `ViewModel`.

A hilt module was added that defines this dependency. In the current
state, this is a bit overkill; but in the future, more services are
going to be used that will be managed by Hilt.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>