app: Reworked ControlUi and its test.
The view model now exposes a Flow with the lookup state to be inline
with other view model classes.

The test no longer uses an Activity. This needed to be changed because
introducing navigation will change the UI flows. It can then no longer
be assumed that the Control UI is displayed directly after starting
the Activity.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>