app: Reworked hierarchy of ServiceDetailsViewModel.
Like has already been done for the view model for the services
overview screen, the abstract base class and the preview subclass have
been dropped. To make this possible, the UI functions no longer
communicate directly with the view model, but expect callback
functions that are used to propagate user interaction.

Again, this is done to simplify the setup required for dependency
injection.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>