app: Fixed an alignment issue with the floating button.
In the services overview screen, some spacing has been added to the
end of the services list. This is needed if there are many services,
to prevent that the floating button blocks the actions of the last
list element. With this space, the list can be further scrolled, so
that the actions of the last list element can be accessed.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>