domain: Added WiFiState enumeration.
This enumeration is going to be used to represent the current state of
the Wi-Fi connection. This is important information, since the app can
only be used to control services if a Wi-Fi connection is available.

The already existing functionality in the `ServerFinder` class is
going to be reworked to follow Clean Android Architecture principles.
This commit is part of this refactoring.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>