domain: Added WiFiStateRepository interface.
This repository interface allows the application to keep track on the
current Wi-Fi connection state.

Being connected to the Wi-Fi is a prerequisite for discovering and
controlling services in the local network. The existing implementation
in the `ServerFinder` class only checks whether the Wi-Fi is
connected, and then ignores this state. In the future, the app is
going to detect changes in the connection state and react accordingly.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>