domain: Added `LookupState` class hierarchy.
These classes allow keeping track on a service discovery operation in
the network. When the URL of a service is not known, it has to be
looked up by sending UDP requests in the network. The state classes
provide some information about this process, so that the UI can be
updated accordingly, until the service is finally located and can be
accessed or the discovery operation failed.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>