build: Disabled classpath scanning of Kotest.
This fixes a warning shown for every test execution and also increases
the startup time for tests significantly. See [1] for reference.

It is certainly not ideal to repeat the required configuration file
for each module, but there does not seem to be an easier way.
Probably, the files can be removed again when version 6.0 is
available.

[1]: https://kotest.io/docs/next/framework/project-config.html#runtime-detection

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>