control: Converted cache to a state flow.
The assumption that the data source is invoked on the main thread does
not hold. It is called by use cases that run on the IO dispatcher.
Therefore, access to the cache for the discovery flows must be
protected. This is achieved by using a `MutableStateFlow` that stores
the current cache map. It is then possible to update the cache in a
thread-safe manner.

Note that the main reason behind this change is to ensure safe
publishing of changes. From the application logic, no real concurrent
updates of the cache are performed; but the cache is read from
different threads.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>