app: `ServicesViewModel` now clears the current service.
When initializing the view model, a `StoreCurrentServiceUseCase` is
now started that clears the name of the current service. That way it
is persisted that currently no control UI is open that needs to be
restored the next time the application is launched. This change needs
to be implemented for other view models as well.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>