app: Changed the state for the control UI.
There are now different sealed classes hierarchies for the discovery
state and the actual UI state. The discovery state is only relevant
when information about the current service could be retrieved and a
discovery could be started. In this case, some more information is
going to be displayed, which has already been prepared in this commit.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>