domain: Added properties to `ServiceDefinition`.
In addition to service discovery, it should also be possible to
directly enter the URL of a service. To support this,
`ServiceDefinition` now supports specifying the mode how the service's
address is determined and a property for the service URL.

So far, the new fields are not evaluated by other parts of the
application.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>