app: Extended `ServiceEditModel` class.
The class now contains logic for handling the new `serviceUrl`
property. There is a new flag which determines whether a URL is
provided for the service or not. Based on this flag, different
validations are performed. This makes sure that the properties for the
selected address mode are valid, while other could have illegal values
that are nevertheless stored.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>