app: Reworked actions in services overview UI.
Instead of displaying multiple icons for each service to trigger
corresponding actions, there is now only a single icon to open a
context menu for a service. From this menu, the desired action can be
triggered.

This approach looks cleaner, it allows extending the number of
available actions, and there is less risk to accidentally click on the
wrong action.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>